% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgpd.R
\name{dgpd}
\alias{dgpd}
\title{The Generalised Pareto Distribution}
\usage{
dgpd(x, scale, shape, shift = 0, shape_tolerance = 1e-10, log = FALSE)
}
\arguments{
\item{x}{Vector of values at which to evaluate the probability density function.}

\item{scale}{Vector of scale parameters, \eqn{\sigma > 0}.}

\item{shape}{Vector of shape parameters, \eqn{\xi \in \mathbb{R}}.}

\item{shift}{Vector of threshold parameters, \eqn{\mu \in \mathbb{R}}.}

\item{shape_tolerance}{Not intended for standard use. Scalar value, such that
when \code{abs(shape) < shape_tolerance}, values are simulated from the limiting
exponential distribution.
@param log Logical; if TRUE then log-density is returned.}
}
\value{
Vector of probability density function values at \code{x}.
}
\description{
These functions provide information about the generalised Pareto distribution:
}
\details{
\itemize{
\item \code{dgpd()} gives the probability density function,
\item \code{pgpd()} gives the cumulative distribution function,
\item \code{qgpd()} gives the quantile function,
\item \code{rgpd()} generates random variates.
}

Any shape values less than \code{shape_tolerance} are drawn from an exponential
distribution using the inverse CDF method. This is mathematically equivalent
to using \code{rexp()} but the seed is not handled in the same way.
}
\examples{
evaluation_points <- c(-1, 0, 0.5, 1, 1.9, 2.1, 5)
dgpd(x = evaluation_points, scale = 1, shape = 0)
dgpd(x = evaluation_points, scale = 1, shape = -0.5)

}
\author{
Zak Varty
}
