% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgpd_mmax_mean.R
\name{dgpd_mmax_mean}
\alias{dgpd_mmax_mean}
\title{The Generalised Pareto Distribution}
\usage{
dgpd_mmax_mean(x, mmax = 2, mean = 1, shift = 0, log = FALSE)
}
\arguments{
\item{x}{Vector of values at which to evaluate the probability density function.}

\item{mmax}{Vector of mmax values without substracting the threshold, \eqn{M_{\max} > mean > shift}.}

\item{mean}{Vector of mean values without substracting the threshold, \eqn{M_{\max} > mean > shift}.}

\item{shift}{Vector of threshold parameters, \eqn{\mu \in \mathbb{R}}.}

\item{log}{Logical; if TRUE then log-density is returned.}
}
\value{
Vector of probability density function values at \code{x}.
}
\description{
Functions for working with the generalised Pareto distribution:
\itemize{
\item \code{dgpd()} gives the probability density function,
\item \code{pgpd()} gives the cumulative distribution function,
\item \code{qgpd()} gives the quantile function,
\item \code{rgpd()} generates random variates.
}

The above functions use the \eqn{(\sigma,\xi)} parametrisation of the
generalised Pareto distribution.

Append "\verb{_nu}" to the function name to use the \eqn{(\nu,\xi)} parametrisation
instead, e.g. \code{dgpd_nu()}.

Append "\verb{_rd}" to the rgpd function to sample observations accounting for
rounding, e.g. \code{rgpd_rd()}.

Append "\verb{_mmax_mean}" to the function name to use the \eqn{(M_{\max}, mean)}
parameterisation, e.g. \code{dgpd_mmax_mean()}. IMPORTANT: This reparameterisation
requires the shape parameter of the original parameterisation to lower than 0,
that is, negative. This is necessary in order to have a finite upper end point.
}
\examples{
evaluation_points <- c(-1, 0, 0.5, 1, 1.9, 2.1, 5)
dgpd_mmax_mean(x = evaluation_points, mmax = 5, mean = 3)
dgpd_mmax_mean(x = evaluation_points, mmax = 10, mean = 4)

}
\author{
Paula Cordero Encinar
}
