% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtgpd.R
\name{dtgpd}
\alias{dtgpd}
\title{The Tapered Generalised Pareto Distribution}
\usage{
dtgpd(
  x,
  scale = 1,
  shape = 0,
  shift = 0,
  scale_taper = 2,
  shape_tolerance = 1e-10,
  log = FALSE
)
}
\arguments{
\item{x}{Vector of values at which to evaluate the probability density function.}

\item{scale}{Vector of scale parameters, \eqn{\sigma > 0}.}

\item{shape}{Vector of shape parameters, \eqn{\xi \in \mathbb{R}}.}

\item{shift}{Vector of threshold parameters, \eqn{\mu \in \mathbb{R}}.}

\item{scale_taper}{Vector of scale parameters for the taper function,
\eqn{\theta > 0}.}

\item{shape_tolerance}{Not intended for standard use. Scalar value, such that
when \code{abs(shape) < shape_tolerance}, values are simulated from the limiting
exponential distribution.}

\item{log}{Logical; if TRUE then log-density is returned.}
}
\value{
Vector of probability density function values at \code{x}.
}
\description{
Functions for working with the tapered generalised Pareto distribution:
\itemize{
\item \code{dtgpd()} gives the probability density function,
\item \code{ptgpd()} gives the cumulative distribution function,
\item \code{rtgpd()} generates random variates.
}

The above functions use the \eqn{(\sigma,\xi)} parametrisation of the
generalised Pareto part of the distribution.

Append "\verb{_rd}" to the rtgpd function to sample observations accounting for
rounding, e.g. \code{rtgpd_rd()}.
}
\details{
Any shape values less than \code{shape_tolerance} are calculated for an exponential
distribution using \code{dexp()}.
}
\examples{
evaluation_points <- c(-1, 0, 0.5, 1, 1.9, 2.1, 5)
dtgpd(x = evaluation_points, scale = 1, shape = 0, scale_taper = 2)
dtgpd(x = evaluation_points, scale = 1, shape = 0.2, scale_taper = 2)
dtgpd(x = evaluation_points, scale = 1, shape = -0.2, scale_taper = 3)

}
\author{
Paula Cordero Encinar
}
