% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earthquakes}
\alias{earthquakes}
\title{List of earthquakes since B.C. around the world from NOAA}
\format{
An object of class \code{data.frame} with 6200 rows and 47 columns.
}
\source{
The Significant Database Earthquake from NOAA
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
earthquakes
}
\description{
Help to access data: https://stackoverflow.com/questions/32964741/accessing-sysdata-rda-within-package-functions
}
\examples{
\dontrun{
 earthquakes
}
}
\keyword{datasets}
