% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Function to clean earthquake data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{Data frame to be cleaned.}
}
\value{
The same data frame with columns ABS_DATE (char column to preserve
BC years), DATE and LOCATION added plus LATITUDE, LONGITUDE,
EQ_PRIMARY and DEATHS columns treated from char to integer class.
}
\description{
This function receives a data frame from The Significant Database Earthquake
from NOAA (https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1) and
return it with new treated columns ABS_DATE, DATE and LOCATION. Also,
turn columns of LATITUDE and LONGITUDE to numeric.
}
\note{
Besides requirements, this function turns columns EQ_PRIMARY and
DEATHS numeric, simplifying their use to aesthetic plots later on.
}
