% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Function to create popup text to leaflet map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Dataframe with earthquake data}
}
\value{
column with HTML formatted data
}
\description{
This function receives a dataframe and return an array of char class
containing a string in a HTML format with informations from three
columns.
}
\examples{
\dontrun{
readr::read_delim("earthquakes.txt", delim = "\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY \%in\% c("CHINA", "USA") & lubridate::year(DATE) >= 2000) \%>\%
mutate(popup_text = eq_create_label(.)) \%>\%
eq_map()
}

}
