% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Functions to clean and plot earthquake data}
\usage{
eq_location_clean(loc_full)
}
\arguments{
\item{loc_full}{Char array to be cleaned}
}
\value{
A cleaned char array
}
\description{
Function to clean location column.
}
\details{
This function receives an array of char containing locations and its
countries joined, returning only the location part. This data is from
'The Significant Database Earthquake' of NOAA.
}
\examples{
\dontrun{readr::read_delim("earthquakes.txt", delim = "\t") \%>\% eq_clean_data()}

}
