% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_map}
\alias{eq_map}
\title{Function to plot earthquakes sites in a map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{Dataframe with earthquake data}

\item{annot_col}{Char array with html data to be plotted in a popup}
}
\value{
Map with earthquakes sites
}
\description{
This function receives a data frame and a column with annotation to be
plotted inside a popup in each earthquake.
}
\examples{
\dontrun{
readr::read_delim("earthquakes.txt", delim = "\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY \%in\% c("CHINA", "USA") & lubridate::year(DATE) >= 2000) \%>\%
mutate(popup_text = eq_create_label(.)) \%>\%
eq_map()
}

}
