% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Geom class to plot earthquakes timelines.}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimeline
}
\arguments{
\item{**x**}{Array of date class}

\item{y}{Array of factor class. Suggested to be the earthquakes country}

\item{colour}{Array of class numeric}

\item{size}{Array of class numeric}

\item{alpha}{Value (from 0 to 1) to control transparency}
}
\value{
Plot layer containing points and segments grob.
}
\description{
This object must receive an array of date class, and optionally could
receive parameters to y (some factor variable, it was proposed earthquake's
country), colour, size and alpha.
}
\keyword{datasets}
