% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llh_gpd_varu.R
\name{llh_gpd_varu}
\alias{llh_gpd_varu}
\title{Log-likelihood of GPD with variable threshold.}
\usage{
llh_gpd_varu(sigxi, u, v, x, negative = FALSE)
}
\arguments{
\item{sigxi}{GPD parameters at threshold u.}

\item{u}{(Low) latent threshold of continuous GPD for which sigxi is given.}

\item{v}{Vector of latent threshold values for each observation.}

\item{x}{Vector of observed values.}

\item{negative}{Logical. Return negative log-likelihood?}
}
\value{
(Negative) log-likelihood of GPD data with variable threshold.
}
\description{
Function to evaluate the (negative) log-likelihood of GPD data
with variable threshold.
}
\examples{
#simulate data
v <- rep(c(1.6,1.2, 1.1), each = 50)
x <- rgpd(n = 150,scale = 1.1, shape = 0.2, shift = v)
llh_gpd_varu(sigxi = c(0.1, -0.01), u = 1, v= v, x=x)
llh_gpd_varu(sigxi = c(0.1, -0.5), u = 1, v= v, x=x)

}
