% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llh_tgpd_varu.R
\name{llh_tgpd_varu}
\alias{llh_tgpd_varu}
\title{Log-likelihood of Tapered GPD with variable threshold.}
\usage{
llh_tgpd_varu(sigxitheta, u, v, x, negative = FALSE)
}
\arguments{
\item{sigxitheta}{Tapered GPD parameters at threshold u.}

\item{u}{(Low) latent threshold of continuous GPD for which sigxi is given.}

\item{v}{Vector of latent threshold values for each observation.}

\item{x}{Vector of observed values.}

\item{negative}{Logical. Return negative log-likelihood?}
}
\value{
(Negative) log-likelihood of tapered GPD data with variable threshold.
}
\description{
Function to evaluate the (negative) log-likelihood of tapered
GPD data with variable threshold.
}
\examples{
#simulate data
v <- rep(c(1.6,1.2, 1.1), each = 50)
x <- rtgpd(n = 150, scale = 1.1, shape = 0.2, shift = v, scale_taper = 2)
llh_tgpd_varu(sigxitheta = c(0.1, -0.01, 0.2), u = 1, v= v, x=x)
llh_tgpd_varu(sigxitheta = c(0.1, -0.5, 0.2), u = 1, v= v, x=x)

}
