% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_gpd_rd_varu.R
\name{mle_gpd_rd_varu}
\alias{mle_gpd_rd_varu}
\title{MLE for rounded GPD data with variable threshold.}
\usage{
mle_gpd_rd_varu(
  sigxi,
  u,
  v,
  x,
  to_nearest = 0.1,
  llh_val = TRUE,
  hessian = FALSE,
  method = "Nelder-Mead",
  maxit = 10000,
  ...
)
}
\arguments{
\item{sigxi}{Initial parameter values for optimisation (latent gpd parameters at threshold u).}

\item{u}{(Low) threshold of latent gpd at which to return optimal parameter values.}

\item{v}{Vector of latent gpd thresholds.}

\item{x}{Vector of rounded gpd observations.}

\item{to_nearest}{Level to which observed values are rounded.}

\item{llh_val}{Logical. Return the log-likelihood value at mle?}

\item{hessian}{Logical. Return numerically estimated hessian at mle?}

\item{method}{The method to be used. See ?optim for more details.}

\item{maxit}{Maximum number of iterations to be passed to optim function.}

\item{...}{additional parameters to be passed to optim function.}
}
\value{
Vector of parameter mles or list containing mles, log-likelihood value and hessian.
}
\description{
This function computes the maximum likelihood estimators for the
for rounded GPD data with variable threshold. In addition, the function can
also return the log-likelihood value as well as a numerical
estimation of the hessian at the MLE.
}
\examples{
#simulate data
v <- rep(c(1.6,1.2, 1.1), each = 50)
x <- rgpd_rd(n = 150,scale = 1.1, shape = 0.2, shift = v, to_nearest = 0.1)
mle_gpd_rd_varu(sigxi = c(1, 0.01), u = 1, v= v, x=x)

}
