% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpd_mmax_mean.R
\name{pgpd_mmax_mean}
\alias{pgpd_mmax_mean}
\title{The Generalised Pareto Distribution}
\usage{
pgpd_mmax_mean(q, mmax = 1, mean = 0, shift = 0)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{mmax}{Vector of mmax values without substracting the threshold, \eqn{M_{\max} > mean > shift}.}

\item{mean}{Vector of mean values without substracting the threshold, \eqn{M_{\max} > mean > shift}.}

\item{shift}{Vector of threshold parameters, \eqn{\mu \in \mathbb{R}}.}
}
\value{
Cumulative distribution function value: \eqn{\Pr(X \leq q)}.
}
\description{
Functions for working with the generalised Pareto distribution:
\itemize{
\item \code{dgpd()} gives the probability density function,
\item \code{pgpd()} gives the cumulative distribution function,
\item \code{qgpd()} gives the quantile function,
\item \code{rgpd()} generates random variates.
}

The above functions use the \eqn{(\sigma,\xi)} parametrisation of the
generalised Pareto distribution.

Append "\verb{_nu}" to the function name to use the \eqn{(\nu,\xi)} parametrisation
instead, e.g. \code{dgpd_nu()}.

Append "\verb{_rd}" to the rgpd function to sample observations accounting for
rounding, e.g. \code{rgpd_rd()}.

Append "\verb{_mmax_mean}" to the function name to use the \eqn{(M_{\max}, mean)}
parameterisation, e.g. \code{dgpd_mmax_mean()}. IMPORTANT: This reparameterisation
requires the shape parameter of the original parameterisation to lower than 0,
that is, negative. This is necessary in order to have a finite upper end point.
}
\examples{
pgpd_mmax_mean(q = seq(-1, 2), mmax = 10, mean = 5)
pgpd_mmax_mean(q = 1, mmax = c(20,10), mean = c(5,6))

}
\author{
Paula Cordero Encinar
}
