% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgpd_mmax_mean.R
\name{qgpd_mmax_mean}
\alias{qgpd_mmax_mean}
\title{The Generalised Pareto Distribution}
\usage{
qgpd_mmax_mean(p, mmax = 2, mean = 1, shift = 0)
}
\arguments{
\item{p}{Vector of probabilities relating to the desired GPD quantiles.}

\item{mmax}{Vector of mmax values without substracting the threshold, \eqn{M_{\max} > mean > shift}.}

\item{mean}{Vector of mean values without substracting the threshold, \eqn{M_{\max} > mean > shift}.}

\item{shift}{Vector of threshold parameters, \eqn{\mu \in \mathbb{R}}.}
}
\value{
Quantile function value: \eqn{x_p} such that \eqn{\Pr(X \leq x_p) = p}.
}
\description{
Functions for working with the generalised Pareto distribution:
\itemize{
\item \code{dgpd()} gives the probability density function,
\item \code{pgpd()} gives the cumulative distribution function,
\item \code{qgpd()} gives the quantile function,
\item \code{rgpd()} generates random variates.
}

The above functions use the \eqn{(\sigma,\xi)} parametrisation of the
generalised Pareto distribution.

Append "\verb{_nu}" to the function name to use the \eqn{(\nu,\xi)} parametrisation
instead, e.g. \code{dgpd_nu()}.

Append "\verb{_rd}" to the rgpd function to sample observations accounting for
rounding, e.g. \code{rgpd_rd()}.

Append "\verb{_mmax_mean}" to the function name to use the \eqn{(M_{\max}, mean)}
parameterisation, e.g. \code{dgpd_mmax_mean()}. IMPORTANT: This reparameterisation
requires the shape parameter of the original parameterisation to lower than 0,
that is, negative. This is necessary in order to have a finite upper end point.
}
\examples{
qgpd_mmax_mean(p = 0.9, mmax = 10, mean = 1, shift = 0)
qgpd_mmax_mean(p = 0.9, mmax = 10, mean = 5, shift = 1)

qgpd_mmax_mean(p = c(0.8, 0.9, 1), mmax = 10, mean = 4, shift = 0)
qgpd_mmax_mean(p = c(0.8, 0.9, 1), mmax = 10, mean = 8, shift = 0)
qgpd_mmax_mean(p = c(0.8, 0.9, 1), mmax = 10, mean = 1, shift = 0)

qgpd_mmax_mean(p = 0.1, mmax = c(11,22,33), mean = 5, shift = c(1,2,3))

}
\author{
Paula Cordero Encinar
}
