% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgpd_nu.R
\name{qgpd_nu}
\alias{qgpd_nu}
\title{The Generalised Pareto Distribution}
\usage{
qgpd_nu(p, scale_alt = 1, shape = 0, shift = 0, shape_tolerance = 1e-10)
}
\arguments{
\item{p}{Vector of probabilities relating to the desired GPD quantiles.}

\item{scale_alt}{Vector of alternative scale parameters, \eqn{\nu = \sigma * (1 + \xi)}.}

\item{shape}{Vector of shape parameters, \eqn{\xi \in \mathbb{R}}.}

\item{shift}{Vector of threshold parameters, \eqn{\mu \in \mathbb{R}}.}

\item{shape_tolerance}{Not intended for standard use. Scalar value, such that
when \code{abs(shape) < shape_tolerance}, values are simulated from the limiting
exponential distribution.}
}
\description{
Functions for working with the generalised Pareto distribution:
\itemize{
\item \code{dgpd()} gives the probability density function,
\item \code{pgpd()} gives the cumulative distribution function,
\item \code{qgpd()} gives the quantile function,
\item \code{rgpd()} generates random variates.
}

The above functions use the \eqn{(\sigma,\xi)} parametrisation of the
generalised Pareto distribution.

Append "\verb{_nu}" to the function name to use the \eqn{(\nu,\xi)} parametrisation
instead, e.g. \code{dgpd_nu()}.
}
\details{
Any shape values less than \code{shape_tolerance} are calculated for an exponential
distribution using \code{dexp()}.
}
\examples{
qgpd_nu(p = 0.9, scale_alt = 1.1, shape = 0.1, shift = 0)
qgpd_nu(p = 0.9, scale_alt = 1.1, shape = 0.1, shift = 1)

qgpd_nu(p = c(0.8, 0.9, 1), scale_alt = 1, shape = 0, shift = 0)
qgpd_nu(p = c(0.8, 0.9, 1), scale_alt = 1.1, shape = 0.1, shift = 0)
qgpd_nu(p = c(0.8, 0.9, 1), scale_alt = 0.9, shape = -0.1, shift = 0)

qgpd_nu(p = 0.1, scale_alt = c(1,2,3), shape = 0, shift = c(1,2,3))

}
\author{
Zak Varty
}
