% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgpd.R
\name{rgpd}
\alias{rgpd}
\title{The Generalised Pareto Distribution}
\usage{
rgpd(n, scale = 1, shape = 0, shift = 0, shape_tolerance = 1e-10)
}
\arguments{
\item{n}{Number of random variates to generate.}

\item{scale}{Vector of scale parameters, sigma > 0.}

\item{shape}{Vector of shape parameters, xi in R.}

\item{shift}{Vector of threshold parameters, mu in R.}

\item{shape_tolerance}{Not intended for standard use. Scalar value, such that when \code{abs(shape) <= shape_tolerance}, values are simulated from an exponential distribution.}
}
\value{
Vector of sampled values from generalised Pareto distribution.
}
\description{
These functions provide information about the generalised Pareto distribution:
}
\details{
\itemize{
\item \code{dgpd()} gives the probability density function,
\item \code{pgpd()} gives the cumulative distribution function,
\item \code{qgpd()} gives the quantile function,
\item \code{rgpd()} generates random variates.
}

Setting xi < 1e-10 draws from an exponential distribution. This is equivalent
to \code{rexp()} but the seed is not handled in the same way.
}
\examples{
rgpd(n = 5, scale = 1, shape = 0, shift = 0, shape_tolerance = 1e-10)

rgpd(n = 5, scale = 1:5, shape = 0.1, shift = 0)
rgpd(n = 5, scale = 1, shape = 0.1 * 1:5, shift = 0)
rgpd(n = 5, scale = 1, shape = 0, shift = 1:5)

}
\author{
Zak Varty
}
