% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgpd_mmax_mean.R
\name{rgpd_mmax_mean}
\alias{rgpd_mmax_mean}
\title{The Generalised Pareto Distribution}
\usage{
rgpd_mmax_mean(n, mmax = 2, mean = 0.5, shift = 0)
}
\arguments{
\item{n}{Number of random variates to generate.}

\item{mmax}{Vector of mmax values without substracting the threshold, \eqn{M_{\max} > mean > shift}.}

\item{mean}{Vector of mean values without substracting the threshold, \eqn{M_{\max} > mean > shift}.}

\item{shift}{Vector of threshold parameters, \eqn{\mu \in \mathbb{R}}.}
}
\value{
Vector of sampled values from generalised Pareto distribution.
}
\description{
Functions for working with the generalised Pareto distribution:
\itemize{
\item \code{dgpd()} gives the probability density function,
\item \code{pgpd()} gives the cumulative distribution function,
\item \code{qgpd()} gives the quantile function,
\item \code{rgpd()} generates random variates.
}

The above functions use the \eqn{(\sigma,\xi)} parametrisation of the
generalised Pareto distribution.

Append "\verb{_nu}" to the function name to use the \eqn{(\nu,\xi)} parametrisation
instead, e.g. \code{dgpd_nu()}.

Append "\verb{_rd}" to the rgpd function to sample observations accounting for
rounding, e.g. \code{rgpd_rd()}.

Append "\verb{_mmax_mean}" to the function name to use the \eqn{(M_{\max}, mean)}
parameterisation, e.g. \code{dgpd_mmax_mean()}. IMPORTANT: This reparameterisation
requires the shape parameter of the original parameterisation to lower than 0,
that is, negative. This is necessary in order to have a finite upper end point.
}
\examples{
rgpd_mmax_mean(n = 5, mmax = 3, mean = 1, shift = 0)
rgpd_mmax_mean(n = 5, mmax = 3:5, mean = 1, shift = 0)
rgpd_mmax_mean(n = 5, mmax = 10, mean = 0.5 * 1:5, shift = 0)
rgpd_mmax_mean(n = 5, mmax = 10, mean = 5, shift = 1:4)

}
\author{
Paula Cordero Encinar
}
