% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgpd_nu.R
\name{rgpd_nu}
\alias{rgpd_nu}
\title{The Generalised Pareto Distribution}
\usage{
rgpd_nu(n, scale_alt = 1, shape = 0, shift = 0, shape_tolerance = 1e-10)
}
\arguments{
\item{n}{Number of random variates to generate.}

\item{scale_alt}{Vector of alternative scale parameters, \eqn{\nu = \sigma * (1 + \xi)}.}

\item{shape}{Vector of shape parameters, \eqn{\xi \in \mathbb{R}}.}

\item{shift}{Vector of threshold parameters, \eqn{\mu \in \mathbb{R}}.}

\item{shape_tolerance}{Not intended for standard use. Scalar value, such that
when \code{abs(shape) < shape_tolerance}, values are simulated from the limiting
exponential distribution.}
}
\value{
Vector of sampled values from generalised Pareto distribution.
}
\description{
These functions provide information about the generalised Pareto distribution:
}
\details{
\itemize{
\item \code{dgpd()} gives the probability density function,
\item \code{pgpd()} gives the cumulative distribution function,
\item \code{qgpd()} gives the quantile function,
\item \code{rgpd()} generates random variates.
}

Any shape values less than \code{shape_tolerance} are drawn from an exponential
distribution using the inverse CDF method. This is mathematically equivalent
to using \code{rexp()} but the seed is not handled in the same way.
}
\examples{
rgpd_nu(n = 5, scale_alt = 1, shape = 0, shift = 0, shape_tolerance = 1e-10)

rgpd_nu(n = 5, scale_alt = 1:5, shape = 0.1, shift = 0)
rgpd_nu(n = 5, scale_alt = 1, shape = 0.1 * 1:5, shift = 0)
rgpd_nu(n = 5, scale_alt = 1, shape = 0, shift = 1:5)

}
\author{
Zak Varty
}
