% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgpd_rd.R
\name{rgpd_rd}
\alias{rgpd_rd}
\title{The Generalised Pareto Distribution}
\usage{
rgpd_rd(
  n,
  scale = 1,
  shape = 0,
  shift = 0,
  shape_tolerance = 1e-10,
  to_nearest = 1,
  shift_latent = NULL
)
}
\arguments{
\item{n}{Number of random variates to generate.}

\item{scale}{Vector of scale parameters, \eqn{\sigma > 0}.}

\item{shape}{Vector of shape parameters, \eqn{\xi \in \mathbb{R}}.}

\item{shift}{Vector of threshold parameters, \eqn{\mu \in \mathbb{R}}.}

\item{shape_tolerance}{Not intended for standard use. Scalar value, such that
when \code{abs(shape) < shape_tolerance}, values are simulated from the limiting
exponential distribution.}

\item{to_nearest}{Precision of resulting observations.}

\item{shift_latent}{Latent shift parameter to account for rounding when sampling.}
}
\description{
Functions for working with the generalised Pareto distribution:
\itemize{
\item \code{dgpd()} gives the probability density function,
\item \code{pgpd()} gives the cumulative distribution function,
\item \code{qgpd()} gives the quantile function,
\item \code{rgpd()} generates random variates.
}

The above functions use the \eqn{(\sigma,\xi)} parametrisation of the
generalised Pareto distribution.

Append "\verb{_nu}" to the function name to use the \eqn{(\nu,\xi)} parametrisation
instead, e.g. \code{dgpd_nu()}.

Append "\verb{_rd}" to the rgpd function to sample observations accounting for
rounding, e.g. \code{rgpd_rd()}.

Append "\verb{_mmax_mean}" to the function name to use the \eqn{(M_{\max}, mean)}
parameterisation, e.g. \code{dgpd_mmax_mean()}.
}
\details{
Any shape values less than \code{shape_tolerance} are drawn from an exponential
distribution using the inverse CDF method. This is mathematically equivalent
to using \code{rexp()} but the seed is not handled in the same way.
}
\examples{
rgpd_rd(n = 5, scale = 1, shape = 0, shift = 0)
rgpd_rd(n = 5, scale = 1:5, shape = 0.1, shift = 0, to_nearest = 0.1)
rgpd_rd(n = 5, scale = 1, shape = 0.1 * 1:5, to_nearest = 0.2)

}
\author{
Paula Cordero Encinar (based on zakvarty/threshold_paper_code)
}
