% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtgpd.R
\name{rtgpd}
\alias{rtgpd}
\title{The Tapered Generalised Pareto Distribution}
\usage{
rtgpd(
  n,
  scale = 1,
  shape = 0,
  shift = 0,
  scale_taper = 2,
  shape_tolerance = 1e-10
)
}
\arguments{
\item{n}{Number of random variates to generate.}

\item{scale}{Vector of scale parameters, \eqn{\sigma > 0}.}

\item{shape}{Vector of shape parameters, \eqn{\xi \in \mathbb{R}}.}

\item{shift}{Vector of threshold parameters, \eqn{\mu \in \mathbb{R}}.}

\item{scale_taper}{Vector of scale parameters for the taper function,
\eqn{\theta > 0}.}

\item{shape_tolerance}{Not intended for standard use. Scalar value, such that
when \code{abs(shape) < shape_tolerance}, values are simulated from the limiting
exponential distribution.}
}
\value{
Vector of sampled values from generalised Pareto distribution.
}
\description{
Functions for working with the tapered generalised Pareto distribution:
\itemize{
\item \code{ptgpd()} gives the cumulative distribution function,
\item \code{rtgpd()} generates random variates.
}

The above functions use the \eqn{(\sigma,\xi)} parametrisation of the
generalised Pareto part of the distribution.

Append "\verb{_rd}" to the rtgpd function to sample observations accounting for
rounding, e.g. \code{rtgpd_rd()}.
}
\details{
Any shape values less than \code{shape_tolerance} are drawn from an exponential
distribution using the inverse CDF method. This is mathematically equivalent
to using \code{rexp()} but the seed is not handled in the same way.
}
\examples{
rtgpd(n = 5, scale = 1, shape = 0, shift = 0, scale_taper = 2)
rtgpd(n = 5, scale = 1:5, shape = 0.1, shift = 0, scale_taper = 10)
rtgpd(n = 5, scale = 1, shape = 0.1 * 1:5, shift = 0, scale_taper = 3)
rtgpd(n = 5, scale = 1, shape = 0, shift = 1:5, scale_taper = 10)
rtgpd(n = 5, scale = 1, shape = 0, shift = 0, scale_taper = 5:10)

}
\author{
Paula Cordero Encinar
}
