% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpd.R
\name{pgpd}
\alias{pgpd}
\title{The Generalised Pareto Distribution}
\usage{
pgpd(q, scale = 1, shape = 0, shift = 0, shape_tolerance = 1e-10)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{scale}{Vector of scale parameters, \eqn{\sigma > 0}.}

\item{shape}{Vector of shape parameters, \eqn{\xi \in \mathbb{R}}.}

\item{shift}{Vector of threshold parameters, \eqn{\mu \in \mathbb{R}}.}

\item{shape_tolerance}{Not intended for standard use. Scalar value, such that
when \code{abs(shape) < shape_tolerance}, values are simulated from the limiting
exponential distribution.}
}
\value{
Cumulative distribution function value: \eqn{\Pr(X \leq q)}.
}
\description{
Functions for working with the generalised Pareto distribution:
\itemize{
\item \code{dgpd()} gives the probability density function,
\item \code{pgpd()} gives the cumulative distribution function,
\item \code{qgpd()} gives the quantile function,
\item \code{rgpd()} generates random variates.
}

The above functions use the \eqn{(\sigma,\xi)} parametrisation of the
generalised Pareto distribution.

Append "\verb{_nu}" to the function name to use the \eqn{(\nu,\xi)} parametrisation
instead, e.g. \code{dgpd_nu()}.

Append "\verb{_rd}" to the rgpd function to sample observations accounting for
rounding, e.g. \code{rgpd_rd()}.

Append "\verb{_mmax_mean}" to the function name to use the \eqn{(M_{\max}, mean)}
parameterisation, e.g. \code{dgpd_mmax_mean()}. IMPORTANT: This reparameterisation
requires the shape parameter of the original parameterisation to lower than 0,
that is, negative. This is necessary in order to have a finite upper end point.
}
\details{
Any shape values less than \code{shape_tolerance} are calculated for an exponential
distribution using \code{pexp()}.
}
\examples{
pgpd(q = seq(-1, 2), shape = 0, scale = 1)
pgpd(q = 1, shape = c(0,-1), scale = c(0.1,1))

}
\author{
Zak Varty (slightly modified by Paula Cordero Encinar)
}
