% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgpd_nu.R
\name{rgpd_nu}
\alias{rgpd_nu}
\title{The Generalised Pareto Distribution}
\usage{
rgpd_nu(n, scale_alt = 1, shape = 0, shift = 0, shape_tolerance = 1e-10)
}
\arguments{
\item{n}{Number of random variates to generate.}

\item{scale_alt}{Vector of alternative scale parameters, nu = sigma * (1 + xi).}

\item{shape}{Vector of shape parameters, xi in R.}

\item{shift}{Vector of threshold parameters, mu in R.}

\item{shape_tolerance}{Not intended for standard use. Scalar value, such that when \code{abs(shape) < shape_tolerance}, values are simulated from an exponential distribution.}
}
\value{
Vector of sampled values from generalised Pareto distribution.
}
\description{
These functions provide information about the generalised Pareto distribution
using the nu parametrisation. For shape parameters greater than xi > -0.5, nu and xi are
approximately orthogonal.
}
\details{
\itemize{
\item \code{dgpd_nu()} gives the probability density function,
\item \code{pgpd_nu()} gives the cumulative distribution function,
\item \code{qgpd_nu()} gives the quantile function,
\item \code{rgpd_nu()} generates random variates.
}

Setting xi < 1e-10 draws from an exponential distribution. This is equivalent
to \code{rexp()} but the seed is not handled in the same way.
}
\examples{
rgpd_nu(n = 5, scale_alt = 1, shape = 0, shift = 0, shape_tolerance = 1e-10)

rgpd_nu(n = 5, scale_alt = 1:5, shape = 0.1, shift = 0)
rgpd_nu(n = 5, scale_alt = 1, shape = 0.1 * 1:5, shift = 0)
rgpd_nu(n = 5, scale_alt = 1, shape = 0, shift = 1:5)

}
\author{
Zak Varty
}
